+++ src/firebench/ros_models/surface_for_vegetation.py
@@ -300,7 +300,6 @@ class Rothermel_SFIRE(RateOfSpreadModel):
             # Spread rate has no dependency on fuel character, only windspeed
             spdms = max(wind, 0.0)
             ros = max(0.03333, 1.2974 * spdms**1.41)  # Spread rate, m/s
-            phiw, phis = np.nan, np.nan
 
         # Default
         return min(ros, 6.0)