+++ src/firebench/stats/__init__.py
@@ -1,0 +1,2 @@
+from .internal.fuel_models import anderson_2015_stats
+from .utils.hist import auto_bins