+++ src/firebench/tools/__init__.py
@@ -44,5 +44,11 @@ from .fuel_models_utils import (
 from .utils import (
     is_scalar_quantity,
     get_value_by_category,
+    calculate_sha256,
 )
 from .raster_to_perimeters import array_to_geopolygons
+from .standard_file_utils import (
+    current_datetime_iso8601,
+    datetime_to_iso8601,
+    read_quantity_from_fb_dataset,
+)