+++ docs/changelog.md
@@ -1,4 +1,4 @@
-# 10. Changelog
+# 11. Changelog
 
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -6,19 +6,39 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [0.7.0]
 ### Added
+- `anderson_2015_stats`: Plot statistics from the Anderson 2015 dataset.
 - `array_to_geopolygons`: Convert an array field into geospatial polygons at a given iso-value, preserving holes.
-- `jaccard_polygon`: Compute the Intersection over Union (IoU), i.e. Jaccard Index, between two fire perimeters described as geospatial polygons.
+- `auto_bins`: Automatically generate histogram bin edges for plotting, based on data range.
+- `CS505_cl`: Compute the half-width of the confidence interval for measurement error of the Campbell Scientific CS505 Fuel Moisture Sensor.
+- `CS506_cl`: Compute the half-width of the confidence interval for measurement error of the Campbell Scientific CS506 Fuel Moisture Sensor.
+- `current_datetime_iso8601`: Get the current datetime as an ISO 8601 formatted string (YYYY-MM-DDTHH:MM[:SS]±HH:MM).
+- `datetime_to_iso8601`: Convert a given datetime to an ISO 8601 formatted string (YYYY-MM-DDTHH:MM[:SS]±HH:MM).
 - `jaccard_binary`: Compute the IoU, i.e. Jaccard Index, between two fire perimeters described as 2D binary masks.
-- `sorensen_dice_polygon`: Compute the Sorensen-Dice index between two fire perimeters described as geospatial polygons.
+- `jaccard_polygon`: Compute the Intersection over Union (IoU), i.e. Jaccard Index, between two fire perimeters described as geospatial polygons.
 - `sorensen_dice_binary`: Compute the Sorensen-Dice index between two fire perimeters described as 2D binary masks.
+- `sorensen_dice_polygon`: Compute the Sorensen-Dice index between two fire perimeters described as geospatial polygons.
+- `read_quantity_from_fb_dataset`: Read a dataset from an HDF5 file, group, or dataset node and return it as a Pint Quantity according to the FireBench I/O standard.
+- `rmse`: Compute the Root Mean Square Error (RMSE) between two arrays, ignoring NaNs.
+- `nmse_range`: Compute the Normalized Mean Square Error (NMSE) between two arrays, using the range of the reference signal as normalization.
+- `nmse_power`: Compute the Normalized Mean Square Error (NMSE) between two arrays, using the product of their mean values as normalization.
 
 ### Documentation
 - Updated developer documentation
+- Refactoring of Benchmarks page. Benchmarks are now sorted by model context instead of metric context.
+- Tags have a colorful badge.
+- Add documentation for FireBench standard file format version 0.1.
+
+### Benchmarks
+- Improve Anderson 2015 Validation benchmark document.
+- Refactoring of `ROS validation using Anderson 2015 dataset` to use the FireBench I/O standard.
 
 ### Miscellaneous
 - Support Python 3.13
 - Support numpy version >= 2.0
 - Dependency to matplotlib > 3.8
+
+### Fix
+- Comma missing in two row of Anderson 2015 dataset.
 
 ## [0.6.1] - 2025 / 05 / 20
 ### Added